Alterando para __init__.py
